Move tree and parseUtils modules under viewer dir
They are only used from viewer/getViewerData so grouping them together
helps in finding out what files reference each other
